refactor: GlobalExceptionHandler에서 예외마다 캐스팅하는 로직 수정
- RuntimeException에는 ErrorCode에 관한 코드가 없어서 각 예외로 직접 캐스팅해야 했으나, ErrorCode를 갖는 CustomException을 만들고 해당 예외를 상속시켜 다형성으로 캐스팅 대체
