+++ src/main/java/subway/exception/CustomException.java
@@ -1,0 +1,15 @@
+package subway.exception;
+
+public class CustomException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+
+    public CustomException(final ErrorCode errorCode, final String message) {
+        super(message);
+        this.errorCode = errorCode;
+    }
+
+    public ErrorCode getErrorCode() {
+        return errorCode;
+    }
+}
