+++ src/main/java/subway/exception/InvalidRequestException.java
@@ -1,15 +1,8 @@
 package subway.exception;
 
-public class InvalidRequestException extends RuntimeException {
-
-    private final ErrorCode errorCode;
+public class InvalidRequestException extends CustomException {
 
     public InvalidRequestException(final ErrorCode errorCode, final String message) {
-        super(message);
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
+        super(errorCode, message);
     }
 }
