+++ src/main/java/subway/exception/LineException.java
@@ -1,15 +1,8 @@
 package subway.exception;
 
-public class LineException extends RuntimeException {
-
-    private final ErrorCode errorCode;
+public class LineException extends CustomException {
 
     public LineException(final ErrorCode errorCode, final String message) {
-        super(message);
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
+        super(errorCode, message);
     }
 }
