+++ src/main/java/subway/exception/SectionException.java
@@ -1,15 +1,8 @@
 package subway.exception;
 
-public class SectionException extends RuntimeException {
-
-    private final ErrorCode errorCode;
+public class SectionException extends CustomException {
 
     public SectionException(final ErrorCode errorCode, final String message) {
-        super(message);
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
+        super(errorCode, message);
     }
 }
